+++ lib/html_parser.dart
@@ -177,17 +177,34 @@ class HtmlRichTextParser extends StatelessWidget {
     "b",
     "i",
     "address",
+    "cite",
+    "var",
     "em",
     "strong",
+    "kbd",
+    "samp",
+    "tt",
     "code",
+    "ins",
     "u",
     "small",
     "abbr",
     "acronym",
+    "mark",
     "ol",
     "ul",
     "blockquote",
+    "del",
+    "s",
+    "strike",
+    "ruby",
+    "rp",
+    "rt",
+    "bdi",
+    "data",
+    "time",
     "span",
+    "big",
   ];
 
   // specialty elements require unique handling
@@ -205,6 +222,7 @@ class HtmlRichTextParser extends StatelessWidget {
     "th",
     "thead",
     "tr",
+    "q",
   ];
 
   // block elements are always rendered with a new
@@ -213,6 +231,7 @@ class HtmlRichTextParser extends StatelessWidget {
   // we simply treat it as a new block level element
   static const _supportedBlockElements = [
     "article",
+    "aside",
     "body",
     "center",
     "dd",
@@ -234,6 +253,8 @@ class HtmlRichTextParser extends StatelessWidget {
     "img",
     "li",
     "main",
+    "nav",
+    "noscript",
     "p",
     "pre",
     "section",
@@ -448,13 +469,19 @@ class HtmlRichTextParser extends StatelessWidget {
             break;
           case "i":
           case "address":
+          case "cite":
+          case "var":
           case "em":
             childStyle =
                 childStyle.merge(TextStyle(fontStyle: FontStyle.italic));
             break;
+          case "kbd":
+          case "samp":
+          case "tt":
           case "code":
             childStyle = childStyle.merge(TextStyle(fontFamily: 'monospace'));
             break;
+          case "ins":
           case "u":
             childStyle = childStyle
                 .merge(TextStyle(decoration: TextDecoration.underline));
@@ -466,8 +493,19 @@ class HtmlRichTextParser extends StatelessWidget {
               decorationStyle: TextDecorationStyle.dotted,
             ));
             break;
+          case "big":
+            childStyle = childStyle.merge(TextStyle(fontSize: 20.0));
+            break;
           case "small":
-            childStyle = childStyle.merge(TextStyle(fontSize: 12.0));
+            childStyle = childStyle.merge(TextStyle(fontSize: 10.0));
+            break;
+          case "mark":
+            childStyle = childStyle.merge(TextStyle(backgroundColor: Colors.yellow, color: Colors.black));
+            break;
+          case "del":
+          case "s":
+          case "strike":
+            childStyle = childStyle.merge(TextStyle(decoration: TextDecoration.lineThrough));
             break;
           case "ol":
             nextContext.indentLevel += 1;
@@ -485,6 +523,12 @@ class HtmlRichTextParser extends StatelessWidget {
             nextContext.indentLevel += 1;
             nextContext.blockType = 'blockquote';
             break;
+          case "ruby":
+          case "rt":
+          case "rp":
+          case "bdi":
+          case "data":
+          case "time":
           case "span":
             //No additional styles
             break;
@@ -620,6 +664,17 @@ class HtmlRichTextParser extends StatelessWidget {
             row.children.add(cell);
             nextContext.parentElement.children.add(row);
             nextContext.parentElement = text.text;
+            break;
+          case "q":
+            if (parseContext.parentElement != null &&
+                parseContext.parentElement is TextSpan) {
+              parseContext.parentElement.children
+                  .add(TextSpan(text: '"', children: []));
+              TextSpan content = TextSpan(text: '', children: []);
+              parseContext.parentElement.children.add(content);
+              parseContext.parentElement.children.add(TextSpan(text: '"', children: []));
+              nextContext.parentElement = content;
+            }
             break;
         }
       }